fix: Instead of exporting OpenApi as a type, export APIInterface
